Split corpus sweeps into per-repository trials
Three corpus checks that each walked the whole corpus in a single test now run one test per corpus repository. The repository list comes from the corpus lockfile, so a failure names the repository instead of the sweep.

The test runner can now run repositories in parallel or filter to one repository. The checks themselves are unchanged.
