+++ Cargo.toml
@@ -38,6 +38,7 @@ dunce = "1.0"
 flate2 = "1.1"
 ignore = "0.4"
 insta = { version = "1.47", features = ["glob", "yaml"] }
+libtest-mimic = "0.8.2"
 memchr = "2.8"
 notify = "8.2"
 percent-encoding = "2.3"
