+++ crates/djls-semantic/Cargo.toml
@@ -27,7 +27,16 @@ djls-testing = { workspace = true }
 
 datatest-stable = { workspace = true }
 insta = { workspace = true }
+libtest-mimic = { workspace = true }
 tempfile = { workspace = true }
+
+[[test]]
+name = "corpus_inheritance"
+harness = false
+
+[[test]]
+name = "corpus_validation"
+harness = false
 
 [[test]]
 name = "mdtest"
